feat(api): add bridge-handler
This is to unify what is required to implement a full bridge between services.

The old way was to listen to each individual event and pray that you listen to all of them and didn't forget one.
