+++ src/instance/minecraft/bridge-handler.ts
@@ -1,0 +1,115 @@
+import type Application from '../../application.js'
+import type {
+  ChatEvent,
+  ClientEvent,
+  CommandEvent,
+  CommandFeedbackEvent,
+  InstanceEvent
+} from '../../common/application-event.js'
+import { ChannelType, EventType, InstanceType } from '../../common/application-event.js'
+import BridgeHandler from '../../common/bridge-handler.js'
+import { antiSpamString } from '../../util/shared-util.js'
+
+import type MinecraftInstance from './minecraft-instance.js'
+
+export default class MinecraftBridgeHandler extends BridgeHandler<MinecraftInstance> {
+  constructor(application: Application, clientInstance: MinecraftInstance) {
+    super(application, clientInstance)
+
+    this.application.on('reconnectSignal', (event) => {
+      // undefined is strictly checked due to api specification
+      if (event.targetInstanceName === undefined || event.targetInstanceName === this.clientInstance.instanceName) {
+        this.clientInstance.logger.log('instance has received restart signal')
+        void this.clientInstance.send(`/gc @Instance restarting...`).then(() => {
+          this.clientInstance.connect()
+        })
+      }
+    })
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  onInstance(event: InstanceEvent): void | Promise<void> {
+    // TODO: implement onInstance on minecraft side
+    // maybe not implement either if it gives better UX
+    return undefined
+  }
+
+  onChat(event: ChatEvent): void | Promise<void> {
+    if (event.instanceName === this.clientInstance.instanceName) return
+
+    if (event.channelType === ChannelType.PUBLIC) {
+      void this.clientInstance.send(this.formatChatMessage('gc', event.username, event.replyUsername, event.message))
+    } else if (event.channelType === ChannelType.OFFICER) {
+      void this.clientInstance.send(this.formatChatMessage('oc', event.username, event.replyUsername, event.message))
+    }
+  }
+
+  onClientEvent(event: ClientEvent): void | Promise<void> {
+    if (event.instanceName === this.clientInstance.instanceName) return
+    if (event.channelType !== ChannelType.PUBLIC) return
+    if (event.removeLater) return
+    if (event.eventType === EventType.BLOCK) return
+    if (event.eventType === EventType.REPEAT) return
+
+    void this.clientInstance.send(`/gc @[${event.instanceName}]: ${event.message}`)
+  }
+
+  onCommand(event: CommandEvent): void | Promise<void> {
+    this.handleCommand(event, false)
+  }
+
+  onCommandFeedback(event: CommandFeedbackEvent): void | Promise<void> {
+    this.handleCommand(event, true)
+  }
+
+  private handleCommand(event: CommandFeedbackEvent, feedback: boolean) {
+    if (
+      event.instanceType === InstanceType.MINECRAFT &&
+      event.instanceName === this.clientInstance.instanceName &&
+      event.alreadyReplied
+    ) {
+      return
+    }
+
+    const finalResponse = `${feedback ? '{f} ' : ''}${event.commandResponse} @${antiSpamString()}`
+    switch (event.channelType) {
+      case ChannelType.PUBLIC: {
+        void this.clientInstance.send(`/gc ${finalResponse}`)
+        break
+      }
+      case ChannelType.OFFICER: {
+        void this.clientInstance.send(`/oc ${finalResponse}`)
+        break
+      }
+      case ChannelType.PRIVATE: {
+        if (event.instanceType !== InstanceType.MINECRAFT || event.instanceName !== this.clientInstance.instanceName)
+          return
+        void this.clientInstance.send(`/msg ${event.username} ${finalResponse}`)
+        break
+      }
+      default: {
+        break
+      }
+    }
+  }
+
+  private formatChatMessage(
+    prefix: string,
+    username: string,
+    replyUsername: string | undefined,
+    message: string
+  ): string {
+    let full = `/${prefix} ${this.clientInstance.bridgePrefix}`
+
+    full += username
+    if (replyUsername != undefined) full += `⇾${replyUsername}`
+    full += ': '
+    full += message
+      .split('\n')
+      .map((s) => s.trim())
+      .join(' ')
+      .trim()
+
+    return full
+  }
+}
